Move files to root directory
