+++ Dockerfile
@@ -10,11 +10,11 @@ ENV PATH /app/node_modules/.bin:$PATH
 # install app dependencies
 COPY package.json ./
 COPY package-lock.json ./
-RUN npm install 
-RUN npm install react-scripts@3.4.1 -g 
+RUN npm install
+RUN npm install react-scripts@3.4.1 -g
 
 # add app
-COPY . ./
+COPY task-manager ./
 
 # start app
 CMD ["npm", "start"]
