+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "task-manager",
+  "name": "Task-Manager-Frontend",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
