+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import * as serviceWorker from './serviceWorker';
+import 'src/index.css';
+import App from 'src/App';
+import * as serviceWorker from 'src/serviceWorker';
 
 ReactDOM.render(
   <React.StrictMode>
